RHINENG-17345: move candlepin client init
