+++ manager/controllers/template_systems_update.go
@@ -22,7 +22,7 @@ import (
 )
 
 var errCandlepin = errors.New("candlepin error")
-var candlepinClient = config.CreateCandlepinClient()
+var candlepinClient = candlepin.CreateCandlepinClient()
 
 type TemplateSystemsUpdateRequest struct {
 	// List of inventory IDs to have templates removed
@@ -197,7 +197,7 @@ func callCandlepin(ctx context.Context, consumer string, request *candlepin.Cons
 	}
 
 	candlepinRespPtr, err := utils.HTTPCallRetry(candlepinFunc,
-		config.CandlepinExpRetries, config.CandlepinRetries, http.StatusServiceUnavailable)
+		candlepin.CandlepinExpRetries, candlepin.CandlepinRetries, http.StatusServiceUnavailable)
 	if err != nil {
 		return nil, errors.Wrap(err, "candlepin /consumers call failed")
 	}
